+++ DRBM/DRBM/DRBMTrainer.h
@@ -7,10 +7,10 @@ class DRBMTrainer
 {
 public:
 	struct {
-		Eigen::VectorXf biasC;
-		Eigen::VectorXf biasD;
-		Eigen::MatrixXf weightXH;
-		Eigen::MatrixXf weightHY;
+		Eigen::VectorXd biasC;
+		Eigen::VectorXd biasD;
+		Eigen::MatrixXd weightXH;
+		Eigen::MatrixXd weightHY;
 	} gradient;
 
 	DRBMOptimizer optimizer;
@@ -20,14 +20,14 @@ public:
 	DRBMTrainer(DRBM & drbm);
 	~DRBMTrainer();
 
-	void train(DRBM & drbm, Eigen::VectorXf & data, int label);
-	double dataMeanXH(DRBM & drbm, Eigen::VectorXf & data, int label, int xindex, int hindex);
-	double dataMeanXH(DRBM & drbm, Eigen::VectorXf & data, int label, int xindex, int hindex, Eigen::MatrixXf & muJK);
-	double dataMeanH(DRBM & drbm, Eigen::VectorXf & data, int label, int hindex);
-	double dataMeanH(DRBM & drbm, Eigen::VectorXf & data, int label, int hindex, Eigen::MatrixXf & muJK);
-	double dataMeanHY(DRBM & drbm, Eigen::VectorXf & data, int label, int hindex, int yindex);
-	double dataMeanHY(DRBM & drbm, Eigen::VectorXf & data, int label, int hindex, int yindex, Eigen::MatrixXf & muJK);
-	double dataMeanY(DRBM & drbm, Eigen::VectorXf & data, int label, int yindex);
-	double dataMeanY(DRBM & drbm, Eigen::VectorXf & data, int label, int yindex, Eigen::MatrixXf & muJK);
+	void train(DRBM & drbm, Eigen::VectorXd & data, int label);
+	double dataMeanXH(DRBM & drbm, Eigen::VectorXd & data, int label, int xindex, int hindex);
+	double dataMeanXH(DRBM & drbm, Eigen::VectorXd & data, int label, int xindex, int hindex, Eigen::MatrixXd & muJK);
+	double dataMeanH(DRBM & drbm, Eigen::VectorXd & data, int label, int hindex);
+	double dataMeanH(DRBM & drbm, Eigen::VectorXd & data, int label, int hindex, Eigen::MatrixXd & muJK);
+	double dataMeanHY(DRBM & drbm, Eigen::VectorXd & data, int label, int hindex, int yindex);
+	double dataMeanHY(DRBM & drbm, Eigen::VectorXd & data, int label, int hindex, int yindex, Eigen::MatrixXd & muJK);
+	double dataMeanY(DRBM & drbm, Eigen::VectorXd & data, int label, int yindex);
+	double dataMeanY(DRBM & drbm, Eigen::VectorXd & data, int label, int yindex, Eigen::MatrixXd & muJK);
 };
